Hold the configuration of a PDA and represent one rule in a PDA’s rulebook
In Chapter 3, we saw that we can simulate a deterministic finite
automaton by keeping track of just one piece of information—the DFA’s
current state—and then using the rulebook to update that information
each time a character is read from the input. But there are two
important things to know about a pushdown automaton at each step of
its computation: what its current state is, and what the current
contents of its stack are. If we use the word configuration to refer
to this combination of a state and a stack, we can talk about a
pushdown automaton moving from one configuration to another as it
reads input characters, which is easier than always having to refer to
the state and stack separately. Viewed this way, a DPDA just has a
current configuration, and the rulebook tells us how to turn the
current configuration into the next configuration each time we read a
character.

Here’s a PDAConfiguration class to hold the configuration of a PDA—a
state and a stack —and a PDARule class to represent one rule in a
PDA’s rulebook.

A rule only applies when the machine’s state, topmost stack
character, and next input character all have the values it expects:

$ irb --simple-prompt
>> require './stack.rb'
>> require './pda_configuration.rb'
>> require './pda_rule.rb'

>> rule = PDARule.new(1, '(', 2, '$', ['b', '$'])
=> #<struct PDARule
   state=1,
   character="(",
   next_state=2,
   pop_character="$",
   push_characters=["b", "$"]
 >
>> configuration = PDAConfiguration.new(1, Stack.new(['$']))
=> #<struct PDAConfiguration state=1, stack=#<Stack ($)>>
>> rule.applies_to?(configuration, '(')
=> true