Follow a PDA rule
For a finite automaton, following a rule just means changing from one
state to another, but a PDA rule updates the stack contents as well as
the state, so PDARule#follow needs to take the machine’s current
configuration as an argument and return the next one.

If we push several characters onto a stack and then pop them off, they
come out in the opposite order:

>> stack = Stack.new(['$']).push('x').push('y').push('z')
=> #<Stack (z)yx$>
>> stack.top
=> "z"
>> stack = stack.pop; stack.top
=> "y"
>> stack = stack.pop; stack.top
=> "x"

PDARule#next_stack anticipates this by reversing the push_characters
array before pushing its characters onto the stack. For example, the
last character in push_characters is the actually the first one to be
pushed onto the stack, so it’ll be the last to be popped off
again. This is just a convenience so that we can read a rule’s
push_characters as the sequence of characters (in “popping order”)
that will be on top of the stack after the rule is applied, without
having to worry about the mechanics of how they get on there.

So, if we have a PDARule that applies to a PDAConfiguration, we can
follow it to find out what the next state and stack will be:

$ irb --simple-prompt
>> require './stack.rb'
>> require './pda_configuration.rb'
>> require './pda_rule.rb'

>> rule = PDARule.new(1, '(', 2, '$', ['b', '$'])
>> configuration = PDAConfiguration.new(1, Stack.new(['$']))

>> rule.follow(configuration)
=> #<struct PDAConfiguration state=2, stack=#<Stack (b)$>>