Support free moves in the simulation
Fine so far, but the rulebook we’re using contains a free move, so
the simulation needs to support free moves before it’ll work
properly. Let’s add a DPDARulebook helper method for dealing with
free moves, similar to the one in NFARulebook.

DPDARulebook#follow_free_moves will repeatedly follow any free moves
that apply to the current configuration, stopping when there are none:

$ irb --simple-prompt
>> require './stack.rb'
>> require './pda_configuration.rb'
>> require './pda_rule.rb'
>> require './dpda_rulebook.rb'

>> rulebook = DPDARulebook.new([
?>  PDARule.new(1, '(', 2, '$', ['b', '$']),
?>  PDARule.new(2, '(', 2, 'b', ['b', 'b']),
?>  PDARule.new(2, ')', 2, 'b', []),
?>  PDARule.new(2, nil, 1, '$', ['$'])
>> ])

>> configuration = PDAConfiguration.new(2, Stack.new(['$']))
=> #<struct PDAConfiguration state=2, stack=#<Stack ($)>>
>> rulebook.follow_free_moves(configuration)
=> #<struct PDAConfiguration state=1, stack=#<Stack ($)>>