Expand DPDA to solve invalid input
There’s one final detail to take care of. Our simulation works
perfectly on inputs that leave the DPDA in a valid state, but it blows
up when the machine gets stuck:

>> dpda_design.accepts?('())')
NoMethodError: undefined method `follow' for nil:NilClass

This happens because DPDARulebook#next_configuration assumes it will
be able to find an applicable rule, so we shouldn’t call it when none
of the rules apply. We’ll fix the problem by modifying
DPDA#read_character to check for a usable rule and, if there isn’t
one, put the DPDA into a special stuck state that it can never move
out of.

Now the DPDA will gracefully become stuck instead of blowing up:

$ irb --simple-prompt
>>  require './stack.rb'
>>  require './pda_configuration.rb'
>>  require './pda_rule.rb'
>>  require './dpda_rulebook.rb'
>>  require './dpda.rb'
>>  require './dpda_design.rb'

>> rulebook = DPDARulebook.new([
>>  PDARule.new(1, '(', 2, '$', ['b', '$']),
>>  PDARule.new(2, '(', 2, 'b', ['b', 'b']),
>>  PDARule.new(2, ')', 2, 'b', []),
>>  PDARule.new(2, nil, 1, '$', ['$'])
>> ])

>> dpda_design = DPDADesign.new(1, '$', [1], rulebook)

>> dpda = DPDA.new(PDAConfiguration.new(1, Stack.new(['$'])), [1], rulebook)
=> #<struct DPDA ...>
>> dpda.read_string('())'); dpda.current_configuration
=> #<struct PDAConfiguration state=#<Object>, stack=#<Stack ($)>>
>> dpda.accepting?
=> false
>> dpda.stuck?
=> true
>> dpda_design.accepts?('())')
=> false