Implement NPDARulebook for holding a nondeterministic collection of PDARules
A nondeterministic machine is more difficult to simulate than a
deterministic one, but we’ve already done the hard work for NFAs in
“Nondeterminism” on page 70, and we can reuse the same ideas for
NPDAs. We need an NPDARulebook for holding a nondeterministic
collection of PDARules, and its implementation is almost exactly the
same as NFARulebook.

In “Nondeterminism” on page 70, we simulated an NFA by keeping track
of a Set of possible states; here we’re simulating an NPDA with a Set
of possible configurations.