Implement a Turing machine tape
This implementation obviously has to store the characters that are
written on the tape, but it also needs to remember the current
position of the tape head so that the simulated machine can read the
current character, write a new character at the current position, and
move the head left and right to reach other positions.

An elegant way of doing this is to split up the tape into three
separate parts—all the characters to the left of the tape head, the
single character directly underneath it, and all the characters to its
right—and store each part separately. This makes it very easy to read
and write the current character, and moving the tape head can be
achieved by shuffling characters between all three parts; moving one
square to the right, for instance, means that the current character
becomes the last character to the left of the head, and the first
character to the right of the head becomes the current one.

Our implementation must also maintain the illusion that the tape is
infinitely long and filled with blank squares, but fortunately, we
don’t need an infinitely large data structure to do this. The only
tape location that can be read at any given moment is the one that’s
beneath the head, so we just need to arrange for a new blank character
to appear there whenever the head is moved beyond the finite number of
nonblank characters already written on the tape. To make this work, we
need to know in advance which character represents a blank square, and
then we can make that character automatically appear underneath the
tape head whenever it moves into an unexplored part of the tape.

This lets us create a tape and read the character beneath the tape head:

$ irb --simple-prompt

>> require './tape.rb'

>> tape = Tape.new(['1', '0', '1'], '1', [], '_')
=> #<Tape 101(1)>
>> tape.middle
=> "1"

>> tape
=> #<Tape 101(1)>
>> tape.move_head_left
=> #<Tape 10(1)1>
>> tape.write('0')
=> #<Tape 101(0)>
>> tape.move_head_right
=> #<Tape 1011(_)>
>> tape.move_head_right.write('0')
=> #<Tape 1011(0)>

Note: Tape, like Stack, is purely functional: writing to the tape and
moving the head are nondestructive operations that return a new Tape
object rather than updating the existing one.