Implement Turing machine configuration and rule
In Chapter 4, we used the word configuration to refer to the
combination of a pushdown automaton’s state and stack, and the same
idea is useful here.

We can say that a Turing machine configuration is the combination of a
state and a tape, and implement Turing machine rules that deal
directly with those configurations.

The unified rule format has five parts:

• The current state of the machine
• The character that must appear at the tape head’s current position
• The next state of the machine
• The character to write at the tape head’s current position
• The direction (left or right) in which to move the head after writing to the tape

A rule only applies when the machine’s current state and the
character currently underneath the tape head match its expectations:

$ irb --simple-prompt

>> require './tape.rb'
>> require './tm_rule.rb'
>> require './tm_configuration.rb'

>> rule = TMRule.new(1, '0', 2, '1', :right)
=> #<struct TMRule
       state=1,
       character="0",
       next_state=2,
       write_character="1",
       direction=:right
   >
>> rule.applies_to?(TMConfiguration.new(1, Tape.new([], '0', [], '_')))
=> true
>> rule.applies_to?(TMConfiguration.new(1, Tape.new([], '1', [], '_')))
=> false
>> rule.applies_to?(TMConfiguration.new(2, Tape.new([], '0', [], '_')))
=> false